requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.4.1
httpx==0.24.1
httpcore==0.17.3
PyMuPDF==1.23.8
python-docx==0.8.11
pytesseract==0.3.10
chromadb==0.4.15
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2
torch==2.0.1
reportlab==4.0.4
weasyprint==60.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.1
requests==2.31.0
faiss-cpu==1.7.4
